+++ packages/core/test/unit/reporters/ci/TravisProvider.spec.ts
@@ -1,73 +1,46 @@
 import { expect } from 'chai';
-import * as sinon from 'sinon';
 
-import * as environmentVariables from '../../../../src/utils/objectUtils';
 import TravisProvider from '../../../../src/reporters/ci/TravisProvider';
+import { EnvironmentVariableStore } from '../../../helpers/EnvironmentVariableStore';
 
-describe('Travis Provider', () => {
-  let getEnvironmentVariables: sinon.SinonStub;
+describe(TravisProvider.name, () => {
+  const env = new EnvironmentVariableStore();
+  let sut: TravisProvider;
 
   beforeEach(() => {
-    getEnvironmentVariables = sinon.stub(environmentVariables, 'getEnvironmentVariable');
+    sut = new TravisProvider();
   });
 
-  describe('isPullRequest()', () => {
-    it('should return false when not building a pull request', () => {
-      getEnvironmentVariables.withArgs('TRAVIS_PULL_REQUEST').returns('false');
-      const travisProvider = new TravisProvider();
-
-      const result = travisProvider.isPullRequest();
-
-      expect(result).to.be.false;
-    });
-
-    it('should return true when building a pull request', () => {
-      getEnvironmentVariables.withArgs('TRAVIS_PULL_REQUEST').returns('42');
-      const travisProvider = new TravisProvider();
-
-      const result = travisProvider.isPullRequest();
-
-      expect(result).to.be.true;
-    });
+  afterEach(() => {
+    env.restore();
   });
 
-  describe('determineBranch()', () => {
-    it('should return the appropriate value', () => {
-      getEnvironmentVariables.withArgs('TRAVIS_BRANCH').returns('master');
-      const travisProvider = new TravisProvider();
-
-      const result = travisProvider.determineBranch();
-
+  describe('determineVersion()', () => {
+    it('should return the current branch', () => {
+      env.set('TRAVIS_BRANCH', 'master');
+      env.unset('TRAVIS_PULL_REQUEST_BRANCH');
+      const result = sut.determineVersion();
       expect(result).to.equal('master');
     });
 
-    it('should return a fall-back value', () => {
-      getEnvironmentVariables.withArgs('TRAVIS_BRANCH').returns('');
-      const travisProvider = new TravisProvider();
-
-      const result = travisProvider.determineBranch();
-
-      expect(result).to.equal('(unknown)');
+    it('should return the PR branch if that is set', () => {
+      env.set('TRAVIS_BRANCH', 'master');
+      env.set('TRAVIS_PULL_REQUEST_BRANCH', 'feat/foo-bar');
+      const result = sut.determineVersion();
+      expect(result).to.equal('feat/foo-bar');
     });
   });
 
-  describe('determineRepository()', () => {
+  describe('determineProject()', () => {
     it('should return the appropriate value', () => {
-      getEnvironmentVariables.withArgs('TRAVIS_REPO_SLUG').returns('stryker/stryker');
-      const travisProvider = new TravisProvider();
+      env.set('TRAVIS_REPO_SLUG', 'stryker/stryker');
+      const result = sut.determineProject();
 
-      const result = travisProvider.determineRepository();
-
-      expect(result).to.equal('stryker/stryker');
+      expect(result).to.equal('github.com/stryker/stryker');
     });
 
-    it('should return a fall-back value', () => {
-      getEnvironmentVariables.withArgs('TRAVIS_REPO_SLUG').returns('');
-      const travisProvider = new TravisProvider();
-
-      const result = travisProvider.determineRepository();
-
-      expect(result).to.equal('(unknown)');
+    it('should return undefined if missing', () => {
+      expect(sut.determineProject()).undefined;
     });
   });
 });