+++ packages/core/test/unit/reporters/dashboard-reporter/DashboardReporter.spec.ts
@@ -1,149 +1,134 @@
 import { mutationTestReportSchema } from '@stryker-mutator/api/report';
 import { testInjector } from '@stryker-mutator/test-helpers';
-import { mutationScoreThresholds } from '@stryker-mutator/test-helpers/src/factory';
+import {
+  mutationTestReportSchemaFileResult,
+  mutationTestReportSchemaMutantResult,
+  mutationTestReportSchemaMutationTestResult
+} from '@stryker-mutator/test-helpers/src/factory';
 import { expect } from 'chai';
 import * as sinon from 'sinon';
+import { ReportType } from '@stryker-mutator/api/core';
 
-import * as ciProvider from '../../../../src/reporters/ci/Provider';
+import { CIProvider } from '../../../../src/reporters/ci/Provider';
 import DashboardReporter from '../../../../src/reporters/dashboard-reporter/DashboardReporter';
 import {
   default as DashboardReporterClient,
-  default as StrykerDashboardClient,
-  StrykerDashboardReport
+  default as StrykerDashboardClient
 } from '../../../../src/reporters/dashboard-reporter/DashboardReporterClient';
 import { dashboardReporterTokens } from '../../../../src/reporters/dashboard-reporter/tokens';
-import * as environmentVariables from '../../../../src/utils/objectUtils';
 import { mock, Mock } from '../../../helpers/producers';
+import { Report } from '../../../../src/reporters/dashboard-reporter/Report';
 
 describe(DashboardReporter.name, () => {
-  let sut: DashboardReporter;
   let dashboardClientMock: Mock<StrykerDashboardClient>;
-  let getEnvironmentVariables: sinon.SinonStub;
-  let determineCiProvider: sinon.SinonStub;
-  const dummyReport: mutationTestReportSchema.MutationTestResult = {
-    files: {},
-    schemaVersion: '1.0',
-    thresholds: mutationScoreThresholds({})
-  };
+  let ciProviderMock: Mock<CIProvider>;
 
   beforeEach(() => {
     dashboardClientMock = mock(StrykerDashboardClient);
-    getEnvironmentVariables = sinon.stub(environmentVariables, 'getEnvironmentVariable');
-    determineCiProvider = sinon.stub(ciProvider, 'determineCIProvider');
+    ciProviderMock = {
+      determineProject: sinon.stub(),
+      determineVersion: sinon.stub()
+    };
   });
 
-  function setupEnvironmentVariables(env?: { ci?: boolean; pullRequest?: boolean; repository?: string; branch?: string; apiKey?: string }) {
-    const { ci, pullRequest, repository, branch, apiKey } = Object.assign(
-      {
-        apiKey: '12345',
-        branch: 'master',
-        ci: true,
-        pullRequest: false,
-        repository: 'stryker-mutator/stryker'
-      },
-      env
-    );
-
-    if (ci) {
-      determineCiProvider.returns({
-        determineBranch: () => branch,
-        determineRepository: () => repository,
-        isPullRequest: () => pullRequest
-      });
-    } else {
-      determineCiProvider.returns(undefined);
-    }
-
-    getEnvironmentVariables.withArgs('STRYKER_DASHBOARD_API_KEY').returns(apiKey);
-    sut = testInjector.injector
+  function createSut(ciProviderOverride: CIProvider | null = ciProviderMock) {
+    return testInjector.injector
       .provideValue(dashboardReporterTokens.dashboardReporterClient, (dashboardClientMock as unknown) as DashboardReporterClient)
+      .provideValue(dashboardReporterTokens.ciProvider, ciProviderOverride)
       .injectClass(DashboardReporter);
   }
 
-  it('should report mutation score to report server', async () => {
+  it('should use the dashboard options if they are available', async () => {
     // Arrange
-    setupEnvironmentVariables();
+    ciProviderMock.determineProject.returns('github.com/foo/bar');
+    ciProviderMock.determineVersion.returns('master');
+    testInjector.options.dashboard.project = 'fooProject';
+    testInjector.options.dashboard.version = 'barVersion';
+    testInjector.options.dashboard.module = 'bazModule';
 
     // Act
-    sut.onMutationTestReportReady({
-      files: {
-        'src/file.js': {
-          language: 'js',
-          mutants: [
-            {
-              id: '1',
-              location: { start: { line: 0, column: 0 }, end: { line: 0, column: 0 } },
-              mutatorName: 'Block',
-              replacement: '{}',
-              status: mutationTestReportSchema.MutantStatus.Killed
-            },
-            {
-              id: '1',
-              location: { start: { line: 0, column: 0 }, end: { line: 0, column: 0 } },
-              mutatorName: 'Block',
-              replacement: '{}',
-              status: mutationTestReportSchema.MutantStatus.Killed
-            },
-            {
-              id: '1',
-              location: { start: { line: 0, column: 0 }, end: { line: 0, column: 0 } },
-              mutatorName: 'Block',
-              replacement: '{}',
-              status: mutationTestReportSchema.MutantStatus.Survived
-            }
-          ],
-          source: 'console.log("hello world!")'
-        }
-      },
-      schemaVersion: '1.0',
-      thresholds: mutationScoreThresholds({})
-    });
+    await act(mutationTestReportSchemaMutationTestResult());
 
     // Assert
-    const report: StrykerDashboardReport = {
-      apiKey: '12345',
-      branch: 'master',
-      mutationScore: 66.66666666666666,
-      repositorySlug: 'github.com/stryker-mutator/stryker'
+    expect(dashboardClientMock.updateReport).calledWithMatch({
+      projectName: 'fooProject',
+      version: 'barVersion',
+      moduleName: 'bazModule'
+    });
+  });
+
+  it('should a update a full report if reportType = "full"', async () => {
+    // Arrange
+    testInjector.options.dashboard.reportType = ReportType.Full;
+    ciProviderMock.determineProject.returns('github.com/foo/bar');
+    ciProviderMock.determineVersion.returns('master');
+    const expectedMutationTestResult = mutationTestReportSchemaMutationTestResult();
+
+    // Act
+    await act(expectedMutationTestResult);
+
+    // Assert
+    expect(dashboardClientMock.updateReport).calledWith({
+      report: expectedMutationTestResult,
+      projectName: 'github.com/foo/bar',
+      version: 'master',
+      moduleName: undefined
+    });
+    expect(testInjector.logger.warn).not.called;
+  });
+
+  it('should a update a mutation score if reportType = "mutationScore', async () => {
+    // Arrange
+    testInjector.options.dashboard.reportType = ReportType.MutationScore;
+    ciProviderMock.determineProject.returns('github.com/foo/bar');
+    ciProviderMock.determineVersion.returns('master');
+    const mutationTestResult = mutationTestReportSchemaMutationTestResult({
+      files: {
+        'a.js': mutationTestReportSchemaFileResult({
+          mutants: [
+            mutationTestReportSchemaMutantResult({ status: mutationTestReportSchema.MutantStatus.Killed }),
+            mutationTestReportSchemaMutantResult({ status: mutationTestReportSchema.MutantStatus.Killed }),
+            mutationTestReportSchemaMutantResult({ status: mutationTestReportSchema.MutantStatus.Killed }),
+            mutationTestReportSchemaMutantResult({ status: mutationTestReportSchema.MutantStatus.Survived })
+          ]
+        })
+      }
+    });
+    const expectedReport: Report = {
+      mutationScore: 75
     };
 
-    expect(dashboardClientMock.postStrykerDashboardReport).to.have.been.calledWith(report);
-    expect(testInjector.logger.warn).to.have.not.been.called;
-  });
-
-  it('should log an info if it is not part of a CI build', async () => {
-    // Arrange
-    setupEnvironmentVariables({ ci: undefined });
-
     // Act
-    sut.onMutationTestReportReady(dummyReport);
+    await act(mutationTestResult);
 
     // Assert
-    expect(dashboardClientMock.postStrykerDashboardReport).to.have.not.been.called;
-    expect(testInjector.logger.info).to.have.been.calledWithMatch('Dashboard report is not sent when not running on a build server');
+    expect(dashboardClientMock.updateReport).calledWith({
+      report: expectedReport,
+      projectName: 'github.com/foo/bar',
+      version: 'master',
+      moduleName: undefined
+    });
+    expect(testInjector.logger.warn).not.called;
   });
 
-  it('should log an info if it is a pull request', async () => {
+  it('should log an info if no settings and no ci provider', async () => {
     // Arrange
-    setupEnvironmentVariables({ pullRequest: true });
+    const sut = createSut(null);
 
     // Act
-    sut.onMutationTestReportReady(dummyReport);
+    sut.onMutationTestReportReady(mutationTestReportSchemaMutationTestResult());
+    await sut.wrapUp();
 
     // Assert
-    expect(dashboardClientMock.postStrykerDashboardReport).to.have.not.been.called;
-    expect(testInjector.logger.info).to.have.been.calledWithMatch('Dashboard report is not sent when building a pull request');
+    expect(dashboardClientMock.updateReport).not.called;
+    expect(testInjector.logger.info).calledWithMatch(
+      'The report was not send to the dashboard. The dashboard.project and/or dashboard.version values were missing and not detected to be running on a build server'
+    );
   });
 
-  it('should log a warning if the Stryker API key is unknown', async () => {
-    // Arrange
-    setupEnvironmentVariables({ apiKey: undefined });
-
-    // Act
-    sut.onMutationTestReportReady(dummyReport);
-
-    // Assert
-    expect(dashboardClientMock.postStrykerDashboardReport).to.have.not.been.called;
-    expect(testInjector.logger.warn).to.have.been.calledWithMatch('Missing environment variable STRYKER_DASHBOARD_API_KEY');
-  });
+  async function act(result: mutationTestReportSchema.MutationTestResult) {
+    const sut = createSut();
+    sut.onMutationTestReportReady(result);
+    await sut.wrapUp();
+  }
 });